Move note/comment feature in a separate viewlet.
