+++ opengever/dossier/tests/test_overview.py
@@ -7,7 +7,11 @@ from lxml.etree import tostring
 from opengever.contact.interfaces import IContactSettings
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_DOSSIER_TEMPLATE_LAYER
 from opengever.core.testing import toggle_feature
+from opengever.dossier.behaviors.dossier import IDossier
 from opengever.testing import FunctionalTestCase
+from plone.protect import createToken
+import json
+import transaction
 
 
 class TestOverview(FunctionalTestCase):
@@ -183,3 +187,83 @@ class TestOverview(FunctionalTestCase):
         self.assertEquals(['Dossier B'], references.text)
         self.assertEquals([dossier_b.absolute_url()],
                           [link.get('href') for link in references])
+
+    @browsing
+    def test_dossier_editlink_for_comments(self, browser):
+        browser.login().visit(self.dossier)
+        # There are both labels (show/hide by css/js)
+        self.assertEquals(['Edit Note', 'Add Note'],
+                          browser.css('#editNoteLink .linkLabel').text)
+
+    @browsing
+    def test_dossier_show_comments_editlink_on_maindossier_only(
+            self, browser):
+        browser.login().visit(self.dossier)
+        comment = browser.css('#editNoteLink')
+        self.assertTrue(comment,
+                        'Expect the edit comment link in dossier byline')
+
+        subdossier = create(Builder('dossier').within(self.dossier))
+        browser.visit(subdossier)
+        comment = browser.css('#editNoteLink')
+        self.assertFalse(comment,
+                         'Expect NO edit comment link on subdossier')
+
+    @browsing
+    def test_dossier_hide_comments_editlink_without_modify_rights(
+            self, browser):
+        self.dossier.manage_permission('Modify portal content',
+                                       roles=[],
+                                       acquire=False)
+        transaction.commit()
+        browser.login().visit(self.dossier)
+        comment = browser.css('#editNoteLink')
+        self.assertFalse(comment,
+                         'Expect NO edit comment link on dossier')
+
+    @browsing
+    def test_dossier_comments_editlink_data(self, browser):
+        browser.login().visit(self.dossier)
+        link = browser.css('#editNoteLink').first
+        dossier_data = IDossier(self.dossier)
+
+        # No comments
+        self.assertEquals(type(json.loads(link.attrib['data-i18n'])),
+                          dict)
+        self.assertEquals('',
+                          link.attrib['data-notecache'])
+        self.assertEquals('Add Note', link.css('.add .linkLabel').first.text)
+
+        # With comments
+        dossier_data.comments = u'This is a comment'
+        transaction.commit()
+
+        browser.open(self.dossier)
+        link = browser.css('#editNoteLink').first
+        self.assertEquals(str(dossier_data.comments),
+                          link.attrib['data-notecache'])
+        self.assertEquals('Edit Note', link.css('.edit .linkLabel').first.text)
+
+    @browsing
+    def test_dossier_save_comments_endpoint(self, browser):
+        payload = '{"comments": "New comment"}'
+        browser.login().visit(self.dossier)
+
+        browser.visit(self.dossier,
+                      view='save_comments',
+                      data={'data': payload,
+                            '_authenticator': createToken()})
+
+        dossier_data = IDossier(self.dossier)
+        self.assertEquals("New comment", dossier_data.comments)
+
+        with self.assertRaises(KeyError):
+            payload = '{"invalidkey": "New comment"}'
+            browser.login().visit(self.dossier,
+                                  view='save_comments',
+                                  data={'data': payload})
+
+        query = {'SearchableText': 'New comment'}
+        result = self.portal.portal_catalog(**query)
+        self.assertEquals(1, len(result), 'Expect one result')
+        self.assertEquals(self.dossier, result[0].getObject())
