+++ opengever/dossier/viewlets/byline.py
@@ -1,15 +1,9 @@
 from ftw.mail.interfaces import IEmailAddress
-from opengever.base import _ as ogbmf
 from opengever.base.viewlets.byline import BylineBase
 from opengever.dossier import _
 from opengever.dossier.base import DOSSIER_STATES_OPEN
 from opengever.dossier.behaviors.dossier import IDossier
-from opengever.dossier.utils import get_main_dossier
 from opengever.ogds.base.actor import Actor
-from plone import api
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from zope.i18n import translate
-import json
 
 
 class BusinessCaseByline(BylineBase):
@@ -26,28 +20,6 @@ class BusinessCaseByline(BylineBase):
     def end(self):
         dossier = IDossier(self.context)
         return self.to_localized_time(dossier.end)
-
-    note_template = ViewPageTemplateFile('templates/note.pt')
-
-    def note(self):
-        dossier = IDossier(self.context)
-        msg_confirm = _(u'note_text_confirm_abord', default=u'Confirm abbord')
-        msg_title_info = ogbmf(u'message_title_info', default=u'Information')
-        msg_title_error = ogbmf(u'message_title_error', default=u'Error')
-        msg_body_info = _(u'message_body_info', default=u'Changes saved')
-        msg_body_error = _(u'message_body_error', default=u'Changes not saved')
-        create_translations = lambda msg: translate(msg, context=self.request)
-
-        translations = json.dumps(
-            {'note_text_confirm_abord': create_translations(msg_confirm),
-             'message_title_info': create_translations(msg_title_info),
-             'message_title_error': create_translations(msg_title_error),
-             'message_body_info': create_translations(msg_body_info),
-             'message_body_error': create_translations(msg_body_error)
-             })
-        comments = '' if not dossier.comments else dossier.comments
-        return self.note_template(note=comments,
-                                  translations=translations)
 
     def mailto_link(self):
         """Displays email-address if the IMailInAddressMarker behavior
@@ -105,16 +77,4 @@ class BusinessCaseByline(BylineBase):
             }
         ]
 
-        can_edit = api.user.has_permission('Modify portal content',
-                                           obj=self.context)
-        is_main_dossier = self.context == get_main_dossier(self.context)
-        if is_main_dossier and can_edit:
-            items += [
-                {
-                    'class': 'note',
-                    'label': _('label_dossier_note', default='Dossiernote'),
-                    'content': self.note(),
-                    'replace': True
-                }
-            ]
         return items
